chore: migrate the existing BIP144 protocol fixture with the transfer fix
